+++ src/interfaces.d.ts
@@ -87,60 +87,7 @@ export type SupportedClassName = string | null | undefined;
 
 export type DeferredVirtualProperties = (inserted: boolean) => VNodeProperties;
 
-/**
- * Animation controls are used to control the web animation that has been applied
- * to a vdom node.
- */
-export interface AnimationControls {
-	play?: boolean;
-	onFinish?: () => void;
-	onCancel?: () => void;
-	reverse?: boolean;
-	cancel?: boolean;
-	finish?: boolean;
-	playbackRate?: number;
-	startTime?: number;
-	currentTime?: number;
-}
-
-export enum AnimationFill { 'none', 'forwards', 'backwards', 'both', 'auto' }
-export enum AnimationDirection { 'normal', 'reverse', 'alternate', 'alternate-reverse' }
-
-/**
- * Animation timing properties passed to a new KeyframeEffect.
- */
-export interface AnimationTimingProperties {
-	duration?: number;
-	delay?: number;
-	direction?: AnimationDirection;
-	easing?: string;
-	endDelay?: number;
-	fill?: AnimationFill;
-	iterations?: number;
-	iterationStart?: number;
-}
-
-/**
- * Animation propertiues that can be passed as vdom property `animate`
- */
-export interface AnimationProperties {
-	id: string;
-	effects: any[];
-	controls?: AnimationControls;
-	timing?: AnimationTimingProperties;
-}
-
-/**
- * Info returned by the `get` function on WebAnimation meta
- */
-export interface AnimationInfo {
-	currentTime: number;
-	playState: 'idle' | 'pending' | 'running' | 'paused' | 'finished';
-	playbackRate: number;
-	startTime: number;
-}
-
-export interface VNodeProperties {
+export interface VirtualDomProperties {
 	/**
 	 * The animation to perform when this node is added to an already existing parent.
 	 * When this value is a string, you must pass a `projectionOptions.transitions` object when creating the
